requirements.txt
# Core dependencies for Lambda automation
pyyaml==6.0.1
boto3==1.28.85
botocore==1.31.85

# Testing
pytest==7.4.3
pytest-cov==4.1.0
moto==4.2.14  # AWS service mocking

# Lambda dependencies (add more as needed)
aws-lambda-powertools==2.25.0

# Development tools
aws-sam-cli==1.102.0
python-terraform==0.10.1  # Terraform Python wrapper for automation

# Utilities
python-dotenv==1.0.0
